there exists a class of error where both a parsing issue and a validation issue can happen for a single occurence of a parameter option. it looks like this:
    # where -o is a required parameter option
    mikey$ ./flarn -o
    expected argument for option '-o'
    --oxygen: required option not provided

that's... sort of true? -o was provided, but incorrectly. this is dumb.

there are several cases i am aware of, all involving required parameter options:

- the option is supplied as the last element of the vector
- a zero-length argument is supplied
- an option-like token is supplied
- the transformer fails on the supplied argument
- parsing fails for the argument slice of an assignment form option

the reason you see both a parsing and a validation error is because both errors actually occurred:

first, the option fails to parse fully. the error is accumulated. no value is placed into the manifest. if there are no other occurrences of the option, or all other occurrences encounter parsing issues, the manifest's value slot for that option will remain empty.

second, when parsing completes and the validator is run, validation detects that the manifest has no data for the option. that generates a second error that is accumulated by the parser.

1. what should a person see when they provide an invocation containing one or more of these issues?

i think it makes the most sense to display the parsing error and not bother telling the user that the option was not provided. it *was* provided, but not correctly. that's what the user cares about.

2. how should CLKit accomplish that?

adding affordance to the manifest is a non-starter. the manifest is a container of successful parsing work, not a container for errors. this should not change.

the validator can't do this (i.e., ignore the applicable constraints) because all it can see is that the manifest does or does not have one or more values for an option. it would still need to know that *a specific* occurrence failed in order to handle the problem correctly for recurrent options. if you wanted to do it in the validator, you would need to provide the equivalent of parsing error information to the validator.

i ain't here to rework parsing and validation in a more significant way, either. they already work well otherwise.

we can fix this in a minimally invasive way by simply eliding validation errors that fall under this umbrella:

if the validation error is one where a required option is not present in the manifest, and we previously saw a parsing error for that option, we know we could not have put anything into the manifest and expect that error. drop the validation error on the floor.